server/requirements.txt
fastapi==0.110.0
pydantic==2.6.3
python-jose==3.3.0
uvicorn==0.27.1
requests==2.31.0
